feat: added param for vehicle - 'finish time'

+++ app/src/main/java/com/yaroslavm87/testtask01/ModelCommands/PassRaceManagerValueRaceTimerToSubscriberModelCommand.java
@@ -19,7 +19,7 @@ public class PassRaceManagerValueRaceTimerToSubscriberModelCommand extends Model
 
     @Override
     public void execute() {
-        Double raceTimerValue = this.sender.getRaceTimerValue();
+        Long raceTimerValue = this.sender.getRaceTimerValue();
         receiver.receiveUpdate(raceTimerValue);
         super.markAsExecuted();
     }
